Last update

+++ SOMOSPIE_pipeline/data-generation-pipeline/load_dems.py
@@ -6,6 +6,7 @@ def load_data(file:str, out_dir:str)->str:
         os.makedirs(out_dir)
     with open(file,'r') as text:
         for line in text:
-            urllib.request.urlretrieve(line, out_dir+line.split("/")[-1])
+            if not os.path.exists(out_dir+line.split("/")[-1]):
+                urllib.request.urlretrieve(line, out_dir+line.split("/")[-1])
     return out_dir
